+++ packages/router/src/index.ts
@@ -1,2 +1,4 @@
+export type { UseObservableOptions } from "@slimr/observable/react"
+export { useObservable } from "@slimr/observable/react"
 export * from "./router-class.js"
 export { Switch } from "./switch.js"