+++ packages/router/src/switch.tsx
@@ -1,4 +1,4 @@
-import React, { memo, useEffect, useState } from "react"
+import React, { memo, useEffect } from "react"
 
 import type { RouterInstance } from "./router-class.js"
 
@@ -6,9 +6,8 @@ import type { RouterInstance } from "./router-class.js"
  * A Switching component, which renders the first matching route
  */
 export const Switch = memo(function Switch({ router }: { router: RouterInstance }) {
-	const [route, setRoute] = useState(router.find(new URL(location.href)))
-	useEffect(() => router.subscribe(setRoute), [])
-	useEffect(router.onLoad, [route])
+	const route = router.route$.use()
+	useEffect(router.onLoad, [route.key])
 	const C = route.component
-	return <C key={route.key} route={route} url={new URL(location.href)} />
+	return <C key={route.key} />
 })